# Add a batch-waitall in MPI_Waitall to reduce cache misses