--- a/src/mpi/pt2pt/bsendutil.c
+++ b/src/mpi/pt2pt/bsendutil.c
@@ -468,7 +468,6 @@ static int MPIR_Bsend_check_active( void )
 
     MPL_DBG_MSG_P(MPIR_DBG_BSEND,TYPICAL,"Checking active starting at %p", active);
     while (active) {
-	MPI_Request r = active->request->handle;
 	int         flag;
 	
 	next_active = active->next;
@@ -483,7 +482,7 @@ static int MPIR_Bsend_check_active( void )
 	    flag = 0;
             /* XXX DJG FIXME-MT should we be checking this? */
 	    if (MPIR_Object_get_ref(active->request) == 1) {
-		mpi_errno = MPIR_Test_impl(&r, &flag, MPI_STATUS_IGNORE );
+                mpi_errno = MPIR_Test_impl(active->request, &flag, MPI_STATUS_IGNORE );
                 if (mpi_errno) MPIR_ERR_POP(mpi_errno);
 	    } else {
 		/* We need to invoke the progress engine in case we 
@@ -495,11 +494,16 @@ static int MPIR_Bsend_check_active( void )
                 if (mpi_errno) MPIR_ERR_POP(mpi_errno);
 	    }
 	} else {
-	    mpi_errno = MPIR_Test_impl( &r, &flag, MPI_STATUS_IGNORE );
+            mpi_errno = MPIR_Test_impl( active->request, &flag, MPI_STATUS_IGNORE );
             if (mpi_errno) MPIR_ERR_POP(mpi_errno);
 	}
 	if (flag) {
 	    /* We're done.  Remove this segment */
+        if (MPIR_Request_is_complete(active->request)) {
+            mpi_errno = MPIR_Request_complete(NULL, active->request, MPI_STATUS_IGNORE,
+                              &flag);
+            if (mpi_errno) MPIR_ERR_POP(mpi_errno);
+        }
 	    MPL_DBG_MSG_P(MPIR_DBG_BSEND,TYPICAL,"Removing segment %p", active);
 	    MPIR_Bsend_free_segment( active );
 	}