--- a/src/mpi/request/testall.c
+++ b/src/mpi/request/testall.c
@@ -38,6 +38,80 @@ int MPI_Testall(int count, MPI_Request array_of_requests[], int *flag,
 int MPIR_Testall_impl(int count, MPI_Request array_of_requests[], int *flag,
                       MPI_Status array_of_statuses[])
 {
+    int i;
+    int mpi_errno = MPI_SUCCESS;
+
+    mpi_errno = MPID_Progress_test();
+    if (mpi_errno != MPI_SUCCESS) goto fn_fail;
+
+    for (i = 0; i < count; i++)
+    {
+        if (request_ptrs[i] != NULL &&
+                request_ptrs[i]->kind == MPIR_REQUEST_KIND__GREQUEST &&
+                request_ptrs[i]->u.ureq.greq_fns->poll_fn != NULL)
+        {
+            mpi_errno = (request_ptrs[i]->u.ureq.greq_fns->poll_fn)(request_ptrs[i]->u.ureq.greq_fns->grequest_extra_state,
+                    &(array_of_statuses[i]));
+            if (mpi_errno != MPI_SUCCESS) goto fn_fail;
+        }
+    }
+
+ fn_exit:
+    return mpi_errno;
+ fn_fail:
+    goto fn_exit;
+}
+
+#endif
+
+#undef FUNCNAME
+#define FUNCNAME MPI_Testall
+#undef FCNAME
+#define FCNAME MPL_QUOTE(FUNCNAME)
+/*@
+    MPI_Testall - Tests for the completion of all previously initiated
+    requests
+
+Input Parameters:
++ count - lists length (integer) 
+- array_of_requests - array of requests (array of handles) 
+
+Output Parameters:
++ flag - True if all requests have completed; false otherwise (logical) 
+- array_of_statuses - array of status objects (array of Status).  May be
+ 'MPI_STATUSES_IGNORE'.
+
+Notes:
+  'flag' is true only if all requests have completed.  Otherwise, flag is
+  false and neither the 'array_of_requests' nor the 'array_of_statuses' is
+  modified.
+
+If one or more of the requests completes with an error, 'MPI_ERR_IN_STATUS' is
+returned.  An error value will be present is elements of 'array_of_status'
+associated with the requests.  Likewise, the 'MPI_ERROR' field in the status
+elements associated with requests that have successfully completed will be
+'MPI_SUCCESS'.  Finally, those requests that have not completed will have a 
+value of 'MPI_ERR_PENDING'.
+
+While it is possible to list a request handle more than once in the
+'array_of_requests', such an action is considered erroneous and may cause the
+program to unexecpectedly terminate or produce incorrect results.
+
+.N ThreadSafe
+
+.N waitstatus
+
+.N Fortran
+
+.N Errors
+.N MPI_SUCCESS
+.N MPI_ERR_IN_STATUS
+.N MPI_ERR_REQUEST
+.N MPI_ERR_ARG
+@*/
+int MPI_Testall(int count, MPI_Request array_of_requests[], int *flag, 
+		MPI_Status array_of_statuses[])
+{
     MPIR_Request * request_ptr_array[MPIR_REQUEST_PTR_ARRAY_SIZE];
     MPIR_Request ** request_ptrs = request_ptr_array;
     MPI_Status * status_ptr;
@@ -47,6 +121,38 @@ int MPIR_Testall_impl(int count, MPI_Request array_of_requests[], int *flag,
     int rc = MPI_SUCCESS;
     int proc_failure = FALSE;
     int mpi_errno = MPI_SUCCESS;
+    MPIR_FUNC_TERSE_STATE_DECL(MPID_STATE_MPI_TESTALL);
+
+    MPIR_ERRTEST_INITIALIZED_ORDIE();
+    
+    MPID_THREAD_CS_ENTER(GLOBAL, MPIR_THREAD_GLOBAL_ALLFUNC_MUTEX);
+    MPIR_FUNC_TERSE_PT2PT_ENTER(MPID_STATE_MPI_TESTALL);
+
+    /* Check the arguments */
+#   ifdef HAVE_ERROR_CHECKING
+    {
+        MPID_BEGIN_ERROR_CHECKS;
+        {
+        int i = 0;
+
+	    MPIR_ERRTEST_COUNT(count, mpi_errno);
+
+	    if (count != 0) {
+		MPIR_ERRTEST_ARGNULL(array_of_requests, "array_of_requests", mpi_errno);
+		/* NOTE: MPI_STATUSES_IGNORE != NULL */
+		MPIR_ERRTEST_ARGNULL(array_of_statuses, "array_of_statuses", mpi_errno);
+	    }
+	    MPIR_ERRTEST_ARGNULL(flag, "flag", mpi_errno);
+
+	    for (i = 0; i < count; i++) {
+		MPIR_ERRTEST_ARRAYREQUEST_OR_NULL(array_of_requests[i], i, mpi_errno);
+	    }
+	}
+        MPID_END_ERROR_CHECKS;
+    }
+#   endif /* HAVE_ERROR_CHECKING */
+
+    /* ... body of routine ...  */
     MPIR_CHKLMEM_DECL(1);
 
     /* Convert MPI request handles to a request object pointers */
@@ -86,14 +192,6 @@ int MPIR_Testall_impl(int count, MPI_Request array_of_requests[], int *flag,
 
     for (i = 0; i < count; i++)
     {
-        if (request_ptrs[i] != NULL &&
-                request_ptrs[i]->kind == MPIR_REQUEST_KIND__GREQUEST &&
-                request_ptrs[i]->u.ureq.greq_fns->poll_fn != NULL)
-        {
-            mpi_errno = (request_ptrs[i]->u.ureq.greq_fns->poll_fn)(request_ptrs[i]->u.ureq.greq_fns->grequest_extra_state,
-                    &(array_of_statuses[i]));
-            if (mpi_errno != MPI_SUCCESS) goto fn_fail;
-        }
         if (request_ptrs[i] != NULL)
         {
             if (MPIR_Request_is_complete(request_ptrs[i]))
@@ -171,107 +269,13 @@ int MPIR_Testall_impl(int count, MPI_Request array_of_requests[], int *flag,
 
     *flag = (n_completed == count) ? TRUE : FALSE;
 
- fn_exit:
+    /* ... end of body of routine ... */
+    
+  fn_exit:
     if (count > MPIR_REQUEST_PTR_ARRAY_SIZE)
     {
         MPIR_CHKLMEM_FREEALL();
     }
-
-    return mpi_errno;
- fn_fail:
-    goto fn_exit;
-}
-
-#endif
-
-#undef FUNCNAME
-#define FUNCNAME MPI_Testall
-#undef FCNAME
-#define FCNAME MPL_QUOTE(FUNCNAME)
-/*@
-    MPI_Testall - Tests for the completion of all previously initiated
-    requests
-
-Input Parameters:
-+ count - lists length (integer) 
-- array_of_requests - array of requests (array of handles) 
-
-Output Parameters:
-+ flag - True if all requests have completed; false otherwise (logical) 
-- array_of_statuses - array of status objects (array of Status).  May be
- 'MPI_STATUSES_IGNORE'.
-
-Notes:
-  'flag' is true only if all requests have completed.  Otherwise, flag is
-  false and neither the 'array_of_requests' nor the 'array_of_statuses' is
-  modified.
-
-If one or more of the requests completes with an error, 'MPI_ERR_IN_STATUS' is
-returned.  An error value will be present is elements of 'array_of_status'
-associated with the requests.  Likewise, the 'MPI_ERROR' field in the status
-elements associated with requests that have successfully completed will be
-'MPI_SUCCESS'.  Finally, those requests that have not completed will have a 
-value of 'MPI_ERR_PENDING'.
-
-While it is possible to list a request handle more than once in the
-'array_of_requests', such an action is considered erroneous and may cause the
-program to unexecpectedly terminate or produce incorrect results.
-
-.N ThreadSafe
-
-.N waitstatus
-
-.N Fortran
-
-.N Errors
-.N MPI_SUCCESS
-.N MPI_ERR_IN_STATUS
-.N MPI_ERR_REQUEST
-.N MPI_ERR_ARG
-@*/
-int MPI_Testall(int count, MPI_Request array_of_requests[], int *flag, 
-		MPI_Status array_of_statuses[])
-{
-    int mpi_errno = MPI_SUCCESS;
-    MPIR_FUNC_TERSE_STATE_DECL(MPID_STATE_MPI_TESTALL);
-
-    MPIR_ERRTEST_INITIALIZED_ORDIE();
-    
-    MPID_THREAD_CS_ENTER(GLOBAL, MPIR_THREAD_GLOBAL_ALLFUNC_MUTEX);
-    MPIR_FUNC_TERSE_PT2PT_ENTER(MPID_STATE_MPI_TESTALL);
-
-    /* Check the arguments */
-#   ifdef HAVE_ERROR_CHECKING
-    {
-        MPID_BEGIN_ERROR_CHECKS;
-        {
-        int i = 0;
-
-	    MPIR_ERRTEST_COUNT(count, mpi_errno);
-
-	    if (count != 0) {
-		MPIR_ERRTEST_ARGNULL(array_of_requests, "array_of_requests", mpi_errno);
-		/* NOTE: MPI_STATUSES_IGNORE != NULL */
-		MPIR_ERRTEST_ARGNULL(array_of_statuses, "array_of_statuses", mpi_errno);
-	    }
-	    MPIR_ERRTEST_ARGNULL(flag, "flag", mpi_errno);
-
-	    for (i = 0; i < count; i++) {
-		MPIR_ERRTEST_ARRAYREQUEST_OR_NULL(array_of_requests[i], i, mpi_errno);
-	    }
-	}
-        MPID_END_ERROR_CHECKS;
-    }
-#   endif /* HAVE_ERROR_CHECKING */
-
-    /* ... body of routine ...  */
-
-    mpi_errno = MPIR_Testall_impl(count, array_of_requests, flag, array_of_statuses);
-    if (mpi_errno != MPI_SUCCESS) goto fn_fail;
-
-    /* ... end of body of routine ... */
-    
-  fn_exit:
     
     MPIR_FUNC_TERSE_PT2PT_EXIT(MPID_STATE_MPI_TESTALL);
     MPID_THREAD_CS_EXIT(GLOBAL, MPIR_THREAD_GLOBAL_ALLFUNC_MUTEX);