--- a/src/mpi/request/wait.c
+++ b/src/mpi/request/wait.c
@@ -174,7 +174,7 @@ int MPI_Wait(MPI_Request *request, MPI_Status *status)
     /* save copy of comm because request will be freed */
     if (request_ptr)
         comm_ptr = request_ptr->comm;
-    mpi_errno = MPIR_Wait_impl(request_ptr, status);
+    mpi_errno = MPID_Wait(request_ptr, status);
     if (mpi_errno) goto fn_fail;
     mpi_errno = MPIR_Request_complete(request, request_ptr, status, &active_flag);
     if (mpi_errno) MPIR_ERR_POP(mpi_errno);