--- a/src/mpi/request/waitall.c
+++ b/src/mpi/request/waitall.c
@@ -147,7 +147,7 @@ int MPIR_Waitall_impl(int count, MPI_Request array_of_requests[],
     }
 
     if (unlikely(disabled_anysource)) {
-        mpi_errno = MPIR_Testall_impl(count, array_of_requests, &disabled_anysource, array_of_statuses);
+        mpi_errno = MPID_Testall(count, request_ptrs, &disabled_anysource, array_of_statuses);
         goto fn_exit;
     }