This is based on PR #2742.
It improves the performance when dealing with large number of requests (by reducing cache misses with the cost of more instructions). It also removes an memory allocation, because the batch size is set to be the available stack buffers for internal requests.

This is WIP, any feedback to improve is appreciated.
(I also fixed the merge conflicts, see only this 3c7411e commit to track changes vs. original PR)
```
window=8192
<2742>
size | MB/s | Messages/s | TOT_INS | L1_DCM
1 | 0.77 | 772205.74 | 1204033.79 | 5069.38
2 | 1.53 | 766035.08 | 2408041.65 | 10116
4 | 2.98 | 746067.09 | 3612049.53 | 15103.13
8 | 6.10 | 761932.77 | 4816057.53 | 20157.45

<this PR>
size | MB/s | Messages/s | TOT_INS | L1_DCM
1 | 0.80 | 798943.71 | 1230194.72 | 1691.25
2 | 1.59 | 792737.86 | 2460361.73 | 3411.09
4 | 3.15 | 786491.78 | 3690528.71 | 5141.10
8 | 6.34 | 792034.11 | 4920695.91 | 6860.15

<master>
size | MB/s | Messages/s | TOT_INS | L1_DCM
1 | 0.78 | 779323.83 | 1163436.04 | 2781.15
2 | 1.56 | 780292.07 | 2326846.47 | 5629.61
4 | 3.15 | 786695.38 | 3490256.54 | 8484.65
8 | 6.34 | 791960.40 | 4653666.75 | 11321.55

window=64 (minimal difference)
<master>

Size | MB/s | Messages/s | TOT_INS | L1_DCM
1 | 0.89 | 886386.69 | 9940.47 | 23.20
2 | 1.81 | 904826.96 | 19853.69 | 46.15
4 | 3.67 | 917355.87 | 29766.89 | 67.70
8 | 7.30 | 911918.77 | 39680.12 | 88.24

<this PR>
Size | MB/s | Messages/s | TOT_INS | L1_DCM
1 | 0.90 | 900083.26 | 10193.47 | 20.20
2 | 1.78 | 890365.15 | 20359.67 | 40.70
4 | 3.68 | 919155.49 | 30525.87 | 62.06
8 | 7.37 | 921270.72 | 40692.07 | 82.91
```