@pavanbalaji 2018-01-28
Thanks, @danghvu.  Can you make this PR rebase-clean on top of `mpich/master`?  It's hard to follow the patches along in this PR.