@pavanbalaji 2018-01-28
@halimamer can you work with @danghvu to clean up this PR?

@danghvu some of @hajimefu's patches seem to be mixing multiple things together into a single patch.  That's discouraged.  Can you split them out into independent pieces?  For example, this patch:

```
MPI: Change MPIR_Wait*_impl semantics

This patch changes MPIR_Wait*_impl (Wait/Waitall/Waitany) semantics in
several ways.
* Now take MPIR_Request * pointer instead of MPI_Request handle
* Do not check for MPI_REQUEST_NULL. It is caller's responsibility
  to ensure the request pointer is valid.
* Do not free the request. They simply make progress until
  the request completes. A caller should then call MPIR_Request_complete
  to free the object and clear the handle.
  (In some minor cases MPIR_Wait*_impl still sets the status,
   for example for generalized request.)
```