@danghvu 2018-01-28
@pavanbalaji I tried to rebase at first but it has a bunch of rebase conflicts which is really painful to fix. That's why I resort to a merge which I only have to fix a few places.

Can you give me an advice on what best to do ? Maybe reworking it with my own commit in a separate PR (moving his code over manually) is better ? Do I need to retain his commit reference / authorship ?