@pavanbalaji 2018-01-28
Explicitly picking code is fine.  If most of the code is copied from @hajimefu's PR, it's best to retain his ownership.  If most of the code is yours, you can ignore that.  Frankly, most of us don't really care about commit authorship, but I don't want to assume that about @hajimefu.  Authorship is mainly to know whom to blame when something goes wrong.  :-)