@danghvu 2018-01-28
Rebased + Reworked to split large commits. It's difficult to split further in those commits that change the `MPIR_*_impl` function to take `MPIR_Request` because other stuff depends on it.