@hajimefu 2018-01-29
@danghvu Thank you for taking a crack on this. As for authorship, I think you can use something like "Co-authored-by: " in the commit message if needed. As @pavanbalaji mentioned, it would help finding who would be responsible for the change later (e.g. ask questions, etc.)