@danghvu 2018-01-29
@hajimefu Thank you, I will do surely do that in the next version of the rework (making a new PR) after gathering some feedbacks from @pavanbalaji