@wesbland 2018-01-30
Yeah, you just have to be careful about how you put it in the commit message. I'd recommend reading the blog post linked above to get it right.