@danghvu 2018-01-30
@pavanbalaji @hajimefu I need to clarify the semantic/idea of this refactoring.

According to the original PR, in the end we are going to create these new functions `MPID_Test*/Wait*` which takes the same arguments as the associated `MPI_Test*/Wait*` series except `MPI_Request` is replaced by `MPIR_Request`.

What is the separation of purposes between the "device" layer and the MPI layer ? For example, should the device set the user flag ? should the device set the status / list of statuses ? Because in many cases I found that the `MPIR_*_impl` series i.e. default implementation of `MPID_*` does not do anything with these arguments.