@halimamer 2018-01-30
test:jenkins/ch3/all