@halimamer 2018-01-31
@danghvu @pavanbalaji , isn't this performance optimization hiding the real problem? The `master` branch does not process requests in batches; it is still allocating the long array on the heap and traverses it in full. If we change `master` to do batched processing without moving the code around, the performance gap might show up again. Have you tested this @danghvu ?