@danghvu 2018-02-01
> Have you tested this @danghvu ?

No I have not tried to batch the original code, I can try later.